// src/app/models/user-profile.ts
export interface UserProfile {
  firstName: string;
  lastName: string;
  email: string;
  phone: string;
  dateOfBirth: string;

  address: string;
  city: string;
  province: string;
  postalCode: string;
  country: string;

  bankName: string;
  accountHolder: string;
  accountNumber: string;
  branchCode: string;
}